pyproject.toml
[project]
name = "lattice"
version = "0.1.0"
description = "A lightweight, composable agent framework for experimentation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Jack Zhu", email = "zhuyijie151@sina.com" },
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "openai>=1.40",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.34"]
chroma = ["chromadb>=0.5"]
all = ["lattice[anthropic,chroma]"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lattice"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
